Payment online

+++ Smq.Web/Assets/client/js/controller/shoppingcart.js
@@ -79,7 +79,19 @@
             if(isValid)
             cart.createOrder();
         });
-        
+        $('input[name="paymentMethod"]').off('click').on('click', function () {
+            if ($(this).val() == 'NL') {
+                $('.boxContent').hide();
+                $('#nganluongContent').show();
+            }
+            else if ($(this).val() == 'ATM_ONLINE') {
+                $('.boxContent').hide();
+                $('#bankContent').show();
+            }
+            else {
+                $('.boxContent').hide();
+            }
+        });
     },
     getLoginUser: function ()
     {
@@ -105,7 +117,8 @@
         CustomerEmail : $('#txtEmail').val(),
         CustomerMobile: $('#txtPhone').val(),
         CustomerMessage : $('#txtMessage').val(),
-        PaymentMethod : "",
+        PaymentMethod: $('input[name="paymentMethod"]:checked').val(),
+        BankCode: $('input[groupname="bankcode"]:checked').prop('id'),
         PaymentStatus : "",
         Status :false
         }
@@ -118,11 +131,20 @@
             },
             success: function (response) {
                 if (response.status) {
-                    $('#divCheckout').hide();
-                    cart.deleteAll();
-                    setTimeout(function () {
-                        $('#cartContent').html("Congratulations on your successful order. We will contact you as soon as possible!");
-                    }, 2000);
+                    if (response.urlCheckout != undefined && response.urlCheckout != '') {
+                        window.location.href = response.urlCheckout;
+                    }
+                    else {
+                        $('#divCheckout').hide();
+                        cart.deleteAll();
+                        setTimeout(function () {
+                            $('#cartContent').html("Congratulations on your successful order. We will contact you as soon as possible!");
+                        }, 2000);
+                    }
+                }
+                else {
+                    $('#divMessage').show();
+                    $('#divMessage').text(response.message);
                 }
             }
         });
